transform StackScriptBase into an HOC
